rocketship: add wrapper for setPermissionCheckHandler to fix notifications
